refactor(worker): extract agent-turn governance and compaction prep
Move Company Brain/model resolution into prepareGovernanceAndModel and
compaction helpers plus portable/remote_v2/pre-inference exits into
prepareCompaction and runPostAgentCompaction.
